move api impl-related code to another module
