+++ components/bsb/timer/background/api/src/commonMain/kotlin/com/flipperdevices/bsb/timer/background/statefactory/TimerStateFactory.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.bsb.timer.background.statefactory
+
+import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
+import com.flipperdevices.bsb.timer.background.model.TimerTimestamp
+
+interface TimerStateFactory {
+    fun create(timestamp: TimerTimestamp): ControlledTimerState
+}
