+++ components/bsb/timer/background/impl/src/commonMain/kotlin/com/flipperdevices/bsb/timer/background/api/statefactory/iteration/IterationBuilder.kt
@@ -1,7 +1,7 @@
-package com.flipperdevices.bsb.timer.background.newstatefactory.iteration
+package com.flipperdevices.bsb.timer.background.api.statefactory.iteration
 
 import com.flipperdevices.bsb.dao.model.TimerSettings
-import com.flipperdevices.bsb.timer.background.newstatefactory.iteration.model.IterationData
+import com.flipperdevices.bsb.timer.background.api.statefactory.iteration.model.IterationData
 import kotlin.time.Duration
 
 interface IterationBuilder {
