+++ components/bsb/timer/background/impl/src/commonTest/kotlin/com/flipperdevices/bsb/timer/background/api/util/ControlledTimerStateFactoryImplTest.kt
@@ -3,15 +3,15 @@ package com.flipperdevices.bsb.timer.background.api.util
 import com.flipperdevices.bsb.dao.model.TimerDuration
 import com.flipperdevices.bsb.dao.model.TimerSettings
 import com.flipperdevices.bsb.dao.model.TimerSettingsId
-import com.flipperdevices.bsb.timer.background.newstatefactory.iteration.model.IterationData
-import com.flipperdevices.bsb.timer.background.newstatefactory.iteration.model.IterationType
+import com.flipperdevices.bsb.timer.background.api.statefactory.iteration.model.IterationData
+import com.flipperdevices.bsb.timer.background.api.statefactory.iteration.model.IterationType
 import com.flipperdevices.bsb.timer.background.statefactory.buildIterationList
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.time.Duration.Companion.hours
 import kotlin.time.Duration.Companion.minutes
 
-class ControlledTimerStateFactoryTest {
+class ControlledTimerStateFactoryImplTest {
     @Test
     fun `GIVEN_standart_iterations_WHEN_build_iterations_THEN_ok`() {
         TimerSettings(
